# target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Article API
  version: 1.0.0
servers:
- url: /
paths:
  /articles:
    get:
      operationId: getArticles
      parameters:
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ArticleDto'
                type: array
          description: A list of articles
      summary: List all articles
      tags:
      - articles
      x-accepts:
      - application/json
      x-tags:
      - tag: articles
    post:
      description: ""
      operationId: createArticle
      requestBody:
        $ref: '#/components/requestBodies/ArticleDto'
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ArticleDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDto'
          description: successful operation
        "405":
          description: Invalid input
      summary: Add a new article to the store
      tags:
      - articles
      x-content-type: application/json
      x-accepts:
      - application/json
      - application/xml
      x-tags:
      - tag: articles
  /articles/{artNo}:
    get:
      operationId: getProductByArtNo
      parameters:
      - explode: false
        in: path
        name: artNo
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDto'
          description: Article found
        "404":
          description: Article not found
      summary: Info for a specific Article
      x-accepts:
      - application/json
components:
  requestBodies:
    ArticleDto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArticleDto'
        application/xml:
          schema:
            $ref: '#/components/schemas/ArticleDto'
      description: Article object that needs to be added to the store
      required: true
  schemas:
    ArticleDto:
      example:
        name: name
        id: 0
        tag: tag
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        tag:
          type: string
      required:
      - name
      type: object